Refactor generator to list into separate helper function
ip.subnet(prefix) returns a generator of IPNetwork objects but we want a list.
Using list(ip.subnet(prefix)) returns a list of IPNetwork objects but we want the string representation here.
So instead we iterate through the generator and build a list of the string representation of each IPNetwork object.