Add IPv6 support
Classful ranges as a name don't make much sense as CIDR ranges predate IPv6 and it's difficult to explain why AWS use these prefixes.
Might want to consider changing the name of the project and the classful/classless terms used in the tests. The actual production code has no mention of these terms.