Allow non strict ranges to be provided
With strict=True (the default), ipaddress.ip_network() throws a ValueError on ranges with host bits set on the supplied address.
Technically the user should provide a proper range without any host bits but we should be able to use a bit of common sense and just fix that for the user.